Make code more robust by introducing loadable()
The original code returned the following warning:

    FutureWarning: elementwise comparison failed; returning scalar
    instead, but in the future will perform elementwise comparison

This comes from the code:

     a in _plots

where a can sometime be a numpy array.  When a is indeed a numpy
array, numpy tells that this comparison should be element-wise, while
pure python tells that this comparison should return scalar.

Currently, the pure python style win and `a in _plots` returns a
scalar as intended.  However, this behavior may change in the future.
Since the elements in "_plots" are all strings.  A clean fix here is
to make sure that "a" is a string become comparing it to the elements
of "_plots".

It is good programming practice to read every warning and make sure
that they are addressed before any problem shows up.