Move Plot into "plot.py" so it is part of the core plotting handlers
The previous method of putting the definition of class Plot in
"ehtplot/plot/__init__.py" makes Plot part of the ehtplot.plot
submodule.  But it really should be viewed as part of the ehtplot
Plot, Panel, and Figure core plotting handlers.  We now move class
Plot in "ehtplot/plot.py" so it's more symmetric compare to
"ehtplot/panel.py" and "ehtplot/figure.py".