Rearchitect: Panel.__call__() returns subaxeses; add Panel.draw()
My new thinking is that Panel.__call__() should be a counterpart of
Panel.__init().  Hence, it should take the same type of arguments,
with the first argument `panelables` replaced by `ax`.  With this new
design, Panel.__call__() cannot accept `*args` to pass to realize
plots.  Therefore, it can only return subaxeses.

We introduce a new Panel.draw() funcion that accepts arbitrary `*args`
and `*kwargs` to pass to Plot.  Similarly, we introduce Plot.draw() to
allow for easy recursion.

Note that we skip call the subpanel touchups for now.